content controller divided changed to small files and route files created. index.ts changed to work properly together
