+++ server/interfaces/tab.interface.ts
@@ -1,0 +1,11 @@
+export interface TabRequestBody {
+  name: string;
+  icon: string;
+  col_one?: string;
+  col_one_b?: string;
+  col_two?: string;
+  col_two_b?: string;
+  col_three?: string;
+  col_three_b?: string;
+  order_no?: number;
+}
